+++ data/trx/ship/games/tr4/scripts/palaces2.lua
@@ -5,4 +5,6 @@ trx.events.on_game_start(function(level)
   trx.items[11].properties.switch_mode = trx.items.SwitchMode.HIDDEN_REACH
   trx.items[124].properties.switch_mode = trx.items.SwitchMode.HIDDEN_PICKUP
   trx.items[126].properties.switch_mode = trx.items.SwitchMode.HIDDEN_PICKUP
+  trx.items[125].properties.pickup_mode = trx.items.PickupMode.HIDDEN
+  trx.items[127].properties.pickup_mode = trx.items.PickupMode.HIDDEN
 end)